+++ src/app/todos/components/main/main.component.ts
@@ -20,11 +20,17 @@ export class MainComponent implements OnInit, OnDestroy {
     editingId: string | null = null;
     tasks: TodoInterface[] = [];
     
+    searchText:string = '';
     private ngUnsubscribe = new Subject<void>();
 
-    constructor(private todoService: TodoService) {}
+    constructor(private todoService: TodoService) {
+        
+    }
 
     ngOnInit(): void {
+
+        
+  
         this.todoService.getTasks().pipe(
             takeUntil(this.ngUnsubscribe)
         ).subscribe(
@@ -42,24 +48,31 @@ export class MainComponent implements OnInit, OnDestroy {
             takeUntil(this.ngUnsubscribe)
         );
 
-        this.visibleTodos$ = combineLatest(
-            this.todoService.todos$,
-            this.todoService.filter$
-        ).pipe(
-            map(([todos, filter]: [TodoInterface[], FilterEnum]) => {
-                console.log("combine =>", todos, filter)
-                if(filter === FilterEnum.active){
-                    console.log("inside active filter");
-                    return todos.filter(todo => !todo.isComplete)
-                }
-                else if(filter === FilterEnum.completed){
-                    console.log("inside completed filter");
-                    return todos.filter(todo => todo.isComplete)
-                }
-                return todos;
-            }),
-            takeUntil(this.ngUnsubscribe)
-        );
+        // this.visibleTodos$ = combineLatest(
+        //     this.todoService.todos$,
+        //     this.todoService.filter$
+        // ).pipe(
+        //     map(([todos, filter]: [TodoInterface[], FilterEnum]) => {
+        //         console.log("combine =>", todos, filter)
+        //         if(filter === FilterEnum.active){
+        //             console.log("inside active filter");
+        //             return todos.filter(todo => !todo.isComplete)
+        //         }
+        //         else if(filter === FilterEnum.completed){
+        //             console.log("inside completed filter");
+        //             return todos.filter(todo => todo.isComplete)
+        //         }
+        //         return todos;
+        //     }),
+        //     takeUntil(this.ngUnsubscribe)
+        // );
+        this.todoService.searchText$
+        .pipe(takeUntil(this.ngUnsubscribe))
+        .subscribe((searchText) => {
+          this.searchText = searchText;
+          this.updateVisibleTodos();
+        });
+        
         
         this.isAllTodosSelected$ = this.todoService.todos$.pipe(
             map((todos) => todos.every((todo) => todo.isComplete)),
@@ -67,6 +80,43 @@ export class MainComponent implements OnInit, OnDestroy {
         );
     }
 
+    
+    updateVisibleTodos(): void {
+        this.visibleTodos$ = combineLatest(
+            this.todoService.todos$,
+            this.todoService.filter$,
+            this.todoService.searchText$
+        ).pipe(
+            map(([todos, filter, searchText]: [TodoInterface[], FilterEnum, string]) => {
+                console.log("SEARCH TEXT====>", searchText);
+                console.log("FILTER====>", filter);
+    
+                let filteredTodos = todos;
+    
+                if (searchText) {
+                    filteredTodos = filteredTodos.filter(todo =>
+                        todo.title.toLowerCase().includes(searchText.toLowerCase())
+                    );
+                }
+    
+                if (filter === FilterEnum.active) {
+                    filteredTodos = filteredTodos.filter(todo => !todo.isComplete);
+                } else if (filter === FilterEnum.completed) {
+                    filteredTodos = filteredTodos.filter(todo => todo.isComplete);
+                }
+    
+                return filteredTodos;
+            }),
+            takeUntil(this.ngUnsubscribe)
+        );
+    
+    
+    }
+    
+      
+
+    
+    
     ngOnDestroy(): void {
         this.ngUnsubscribe.next();
         this.ngUnsubscribe.complete();
@@ -74,6 +124,7 @@ export class MainComponent implements OnInit, OnDestroy {
 
     toggleAllTodos(event : Event): void {
         const target = event.target as HTMLInputElement
+        console.log("Target===>",target);
         this.todoService.toggleAll(target.checked);
     }
 
@@ -83,7 +134,15 @@ export class MainComponent implements OnInit, OnDestroy {
     }
     
     removeTodoHandler(taskId: string): void {
-        this.todoService.removeTodo(taskId);
+        const confirmstatus = window.confirm('Are you sure you want to delete?');
+        if(confirmstatus){
+            this.todoService.removeTodo(taskId);
+        }
+        else{
+            console.log("cancelled deleting the task");
+            
+        }
+        
     }
 
     toggleTodoHandler(taskId: string): void {
